Refactor testing structure and remove deprecated tests
- Deleted integration tests from `tests/shell_integration_tests.rs` and unit tests from `tests/shell_tests.rs` as part of the transition to a modular shell architecture.
- Updated validation report timestamp to reflect recent changes.
- Created a comprehensive Code Issues Analysis document detailing identified issues and resolutions during the refactor.
- Established a new Testing Strategy document outlining the current testing infrastructure and available test scripts post-refactor.

+++ src/shell/commands/memory.rs
@@ -60,7 +60,7 @@ pub fn handle_memory_allocate(uart: &Uart, memory_manager: &mut MemoryManager) {
     match memory_manager.allocate_block() {
         Some(address) => {
             uart.puts("✓ Memory block allocated at: 0x");
-            print_hex(uart, address as u32);
+            print_hex(uart, address);
             uart.puts("\r\n");
         }
         None => {
